feat: Gọi API màn UserDetail
