+++ app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileViewModel.kt
@@ -46,5 +46,10 @@ class ProfileViewModel @Inject constructor(
         }
     }
 
+    fun getUserData(): Pair<String, String> {
+        return Pair(_uiState.value.username, _uiState.value.userAvatar)
+    }
+
+
     // TODO: Load data from server
 }
