+++ app/src/main/java/com/pwhs/quickmem/presentation/app/user_detail/UserDetailUiAction.kt
@@ -1,4 +1,6 @@
 package com.pwhs.quickmem.presentation.app.user_detail
 
 sealed class UserDetailUiAction {
+    data object Refresh : UserDetailUiAction()
+    data class ShowError(val message: String) : UserDetailUiAction()
 }
