+++ app/src/main/java/com/pwhs/quickmem/presentation/app/user_detail/UserDetailUiEvent.kt
@@ -1,6 +1,6 @@
 package com.pwhs.quickmem.presentation.app.user_detail
 
 sealed class UserDetailUiEvent {
-    data object Refresh : UserDetailUiEvent()
     data class ShowError(val message: String) : UserDetailUiEvent()
+    data object Refresh : UserDetailUiEvent()
 }
