+++ app/src/main/java/com/pwhs/quickmem/presentation/app/user_detail/UserDetailViewModel.kt
@@ -3,25 +3,25 @@ package com.pwhs.quickmem.presentation.app.user_detail
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.pwhs.quickmem.core.datastore.TokenManager
 import com.pwhs.quickmem.core.utils.Resources
 import com.pwhs.quickmem.domain.repository.UserRepository
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class UserDetailViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val tokenManager: TokenManager
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(UserDetailUiState())
-    val uiState = _uiState.asStateFlow()
+    val uiState: StateFlow<UserDetailUiState> = _uiState
 
-    private val _uiEvent = Channel<UserDetailUiEvent>()
-    val uiEvent = _uiEvent.receiveAsFlow()
+    private val _uiEvent = Channel<UserDetailUiEvent>(Channel.BUFFERED)
+    val uiEvent: Flow<UserDetailUiEvent> = _uiEvent.receiveAsFlow()
 
     init {
         val userId = savedStateHandle.get<String>("userId") ?: ""
@@ -30,35 +30,40 @@ class UserDetailViewModel @Inject constructor(
 
     private fun loadUserDetails(userId: String) {
         viewModelScope.launch {
-            val token = "token"
-            userRepository.getUserById(userId, token).collect { resource ->
-                when (resource) {
-                    is Resources.Loading -> {
-                        _uiState.value = _uiState.value.copy(isLoading = true, errorMessage = null)
-                    }
-                    is Resources.Success -> {
-                        val user = resource.data
-                        if (user != null) {
-                            _uiState.value = _uiState.value.copy(
-                                isLoading = false,
-                                userId = user.id,
-                                userName = user.username,
-                                userEmail = "",
-                                avatarUrl = user.avatarUrl,
-                                errorMessage = null
-                            )
-                        } else {
-                            sendUiEvent(UserDetailUiEvent.ShowError("User data is null"))
+            tokenManager.accessToken
+                .onStart {
+                    _uiState.value = _uiState.value.copy(isLoading = true, errorMessage = null)
+                }
+                .collect { token ->
+                    if (!token.isNullOrEmpty()) {
+                        userRepository.getUserById(userId, token).collect { resource ->
+                            when (resource) {
+                                is Resources.Loading -> {
+                                    _uiState.value = _uiState.value.copy(isLoading = true, errorMessage = null)
+                                }
+                                is Resources.Success -> {
+                                    val user = resource.data
+                                    _uiState.value = _uiState.value.copy(
+                                        isLoading = false,
+                                        userName = user?.username.orEmpty(),
+                                        avatarUrl = user?.avatarUrl.orEmpty(),
+                                        errorMessage = null
+                                    )
+                                }
+                                is Resources.Error -> {
+                                    _uiState.value = _uiState.value.copy(isLoading = false)
+                                    sendUiEvent(UserDetailUiEvent.ShowError(resource.message ?: "An unknown error occurred"))
+                                }
+                            }
                         }
-                    }
-                    is Resources.Error -> {
+                    } else {
                         _uiState.value = _uiState.value.copy(isLoading = false)
-                        sendUiEvent(UserDetailUiEvent.ShowError(resource.message ?: "An unknown error occurred"))
+                        sendUiEvent(UserDetailUiEvent.ShowError("Token is missing or invalid"))
                     }
                 }
-            }
         }
     }
+
 
     private fun sendUiEvent(event: UserDetailUiEvent) {
         viewModelScope.launch {
