--- app/src/main/java/com/pwhs/quickmem/presentation/app/user_detail/component/UserDetail.kt
@@ -1,118 +1,0 @@
-package com.pwhs.quickmem.presentation.app.user_detail.component
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.*
-import androidx.compose.material3.TabRowDefaults.SecondaryIndicator
-import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
-import com.pwhs.quickmem.R
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun UserDetail(
-    modifier: Modifier = Modifier,
-    isLoading: Boolean = false,
-    userName: String = "User Name",
-    avatarUrl: String = "",
-    onBackClick: () -> Unit
-) {
-    var tabIndex by remember { mutableIntStateOf(0) }
-    val tabTitles = listOf("STUDY SET", "CLASS", "FOLDER")
-
-    Scaffold(
-        modifier = modifier.fillMaxSize(),
-        topBar = {
-            TopAppBar(
-                title = { },
-                navigationIcon = {
-                    IconButton(onClick = { onBackClick() }) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "Back"
-                        )
-                    }
-                },
-                actions = {}
-            )
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Top
-        ) {
-            if (isLoading) {
-                CircularProgressIndicator(
-                    modifier = Modifier
-                        .padding(top = 16.dp)
-                        .size(50.dp),
-                    color = MaterialTheme.colorScheme.primary
-                )
-            } else {
-                AsyncImage(
-                    model = ImageRequest.Builder(LocalContext.current)
-                        .data(avatarUrl.ifEmpty { null })
-                        .placeholder(R.drawable.default_avatar)
-                        .error(R.drawable.default_avatar)
-                        .build(),
-                    contentDescription = "User Avatar",
-                    modifier = Modifier
-                        .size(100.dp)
-                        .clip(CircleShape),
-                    contentScale = ContentScale.Crop
-                )
-
-                Text(
-                    text = userName,
-                    style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold),
-                    modifier = Modifier.padding(top = 8.dp)
-                )
-            }
-
-            TabRow(
-                selectedTabIndex = tabIndex,
-                indicator = { tabPositions ->
-                    SecondaryIndicator(
-                        Modifier.tabIndicatorOffset(tabPositions[tabIndex]),
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                },
-                modifier = Modifier.padding(top = 16.dp)
-            ) {
-                tabTitles.forEachIndexed { index, title ->
-                    Tab(
-                        text = { Text(title) },
-                        selected = tabIndex == index,
-                        onClick = { tabIndex = index }
-                    )
-                }
-            }
-
-            when (tabIndex) {
-                0 -> {
-                    Text(text = "Content for STUDY SET", modifier = Modifier.padding(16.dp))
-                }
-                1 -> {
-                    Text(text = "Content for CLASS", modifier = Modifier.padding(16.dp))
-                }
-                2 -> {
-                    Text(text = "Content for FOLDER", modifier = Modifier.padding(16.dp))
-                }
-            }
-        }
-    }
-}
