+++ app/src/main/java/com/pwhs/quickmem/presentation/app/user_detail/component/UserDetailTabEnum.kt
@@ -1,0 +1,7 @@
+package com.pwhs.quickmem.presentation.app.user_detail.component
+
+enum class UserDetailTabEnum(val index: Int) {
+    STUDY_SET(0),
+    CLASS(1),
+    FOLDER(2)
+}
